+++ include/sterm/terminal.hpp
@@ -65,6 +65,7 @@ namespace sterm {
       void          spawn_child(std::string command);
 
       void          reload_configuration();
+      void          reset_terminal(gboolean full);
 
       std::string   get_window_title();
       std::string   get_text();